Reogranized TestUtils -> AbstractPolicyTest
