+++ driver-core/src/test/java/com/datastax/driver/core/TestUtils.java
@@ -40,27 +40,6 @@ public abstract class TestUtils {
 
     public static final String INSERT_FORMAT = "INSERT INTO %s (k, t, i, f) VALUES ('%s', '%s', %d, %f)";
     public static final String SELECT_ALL_FORMAT = "SELECT * FROM %s";
-    public static final String SELECT_WHERE_FORMAT = "SELECT * FROM %s WHERE %s";
-
-    public static void createSchema(Session session) {
-        createSchema(session, 1);
-    }
-
-    public static void createSchema(Session session, int replicationFactor) {
-        session.execute(String.format(CREATE_KEYSPACE_SIMPLE_FORMAT, SIMPLE_KEYSPACE, replicationFactor));
-        session.execute("USE " + SIMPLE_KEYSPACE);
-        session.execute(String.format("CREATE TABLE %s (k int PRIMARY KEY, i int)", SIMPLE_TABLE));
-    }
-
-    public static void createMultiDCSchema(Session session) {
-        createMultiDCSchema(session, 1, 1);
-    }
-
-    public static void createMultiDCSchema(Session session, int dc1RF, int dc2RF) {
-        session.execute(String.format(CREATE_KEYSPACE_GENERIC_FORMAT, SIMPLE_KEYSPACE, "NetworkTopologyStrategy", String.format("'dc1' : 1, 'dc2' : 1", dc1RF, dc2RF)));
-        session.execute("USE " + SIMPLE_KEYSPACE);
-        session.execute(String.format("CREATE TABLE %s (k int PRIMARY KEY, i int)", SIMPLE_TABLE));
-    }
 
     public static BoundStatement setBoundValue(BoundStatement bs, String name, DataType type, Object value) {
         switch (type.getName()) {
